+++ app/src/main/AndroidManifest.xml
@@ -35,45 +35,45 @@
         android:theme="@style/Theme.AppCompat.Light.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".MapMemoActivity"
+            android:name=".ui.mapmemo.MapMemoActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".DiarySelectedActivity"
+            android:name=".ui.diary.DiarySelectedActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".TodolistActivity"
+            android:name=".ui.todolist.TodolistActivity"
             android:exported="false" />
         <activity
-            android:name=".DiaryWritingPageActivity"
+            android:name=".ui.diary.DiaryWritingPageActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".DiaryListActivity"
+            android:name=".ui.diary.DiaryListActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".RingingAlarmActivity"
+            android:name=".ui.alarm.RingingAlarmActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".AlarmActivity"
+            android:name=".ui.alarm.AlarmActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
@@ -93,7 +93,7 @@
                 android:value="" />
         </activity>
 
-        <receiver android:name=".AlarmReceiver" />
+        <receiver android:name=".ui.alarm.AlarmReceiver" />
 
         <provider
             android:name="androidx.core.content.FileProvider"
