+++ app/src/main/java/smu/team3_orda_diary/model/OnePageDiary.java
@@ -1,4 +1,4 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.model;
 
 import android.graphics.Color;
