+++ app/src/main/java/smu/team3_orda_diary/model/TodoItem.java
@@ -1,4 +1,4 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.model;
 
 public class TodoItem {
