+++ app/src/main/java/smu/team3_orda_diary/ui/alarm/AlarmFragment.java
@@ -1,4 +1,4 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.ui.alarm;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import smu.team3_orda_diary.R;
 
 /**
  * A simple {@link Fragment} subclass.
