+++ app/src/main/java/smu/team3_orda_diary/ui/diary/DiarySelectedActivity.java
@@ -1,8 +1,8 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.ui.diary;
 
 import androidx.appcompat.app.AppCompatActivity;
 // 일기들이 저장되어 있는 diaryList에 접근하는 용도
-import static smu.team3_orda_diary.DiaryListActivity.diaryList;
+import static smu.team3_orda_diary.ui.diary.DiaryListActivity.diaryList;
 // 텍스트뷰의 스크롤을 위한 것
 import android.text.method.ScrollingMovementMethod;
 // 그 외 기본적인 필요한 것들
@@ -12,6 +12,9 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import smu.team3_orda_diary.model.OnePageDiary;
+import smu.team3_orda_diary.R;
 
 public class DiarySelectedActivity extends AppCompatActivity {
     int clickPosition;
