+++ app/src/main/java/smu/team3_orda_diary/ui/home/HomeFragment.java
@@ -1,4 +1,4 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.ui.home;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import smu.team3_orda_diary.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link DiaryListFragment#newInstance} factory method to
+ * Use the {@link HomeFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class DiaryListFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class DiaryListFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public DiaryListFragment() {
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class DiaryListFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment DiaryListFragment.
+     * @return A new instance of fragment HomeFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static DiaryListFragment newInstance(String param1, String param2) {
-        DiaryListFragment fragment = new DiaryListFragment();
+    public static HomeFragment newInstance(String param1, String param2) {
+        HomeFragment fragment = new HomeFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class DiaryListFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_diary_list, container, false);
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 }
