+++ app/src/main/java/smu/team3_orda_diary/ui/todolist/ToDoListFragment.java
@@ -1,4 +1,4 @@
-package smu.team3_orda_diary;
+package smu.team3_orda_diary.ui.todolist;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -20,7 +20,10 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Locale;
 
+import smu.team3_orda_diary.database.DBHelper;
+import smu.team3_orda_diary.R;
 import smu.team3_orda_diary.databinding.FragmentToDoListBinding;
+import smu.team3_orda_diary.model.TodoItem;
 
 public class ToDoListFragment extends Fragment {
