+++ app/src/main/res/layout/activity_diary_writing_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".DiaryWritingPageActivity">
+    tools:context=".ui.diary.DiaryWritingPageActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
