+++ app/src/main/res/layout/activity_map_memo.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".MapMemoActivity">
+    tools:context=".ui.mapmemo.MapMemoActivity">
 
     <TextView
         android:layout_width="match_parent"
