+++ app/src/main/res/layout/activity_ringing_alarm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RingingAlarmActivity">
+    tools:context=".ui.alarm.RingingAlarmActivity">
 
     <ImageView
         android:id="@+id/imageView4"
