+++ app/src/main/res/layout/activity_todolist.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TodolistActivity">
+    tools:context=".ui.todolist.TodolistActivity">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/todo_btn_write"
