+++ app/src/main/res/layout/fragment_map_memo.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapMemoFragment">
+    tools:context=".ui.mapmemo.MapMemoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
