+++ app/src/main/res/layout/one_page_dirary_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    tools:context=".DiaryWritingPageActivity">
+    tools:context=".ui.diary.DiaryWritingPageActivity">
 
     <ImageView
         android:id="@+id/imageView"
